requirements.txt
altgraph==0.17.4
attrs==22.1.0
autogui==0.1.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
clr-loader==0.2.6
colorama==0.4.5
comtypes==1.2.1
contextlib2==21.6.0
cryptography==38.0.1
cycler==0.11.0
deprecation==2.1.0
EasyProcess==1.1
entrypoint2==1.1
execnet==1.9.0
fonttools==4.33.3
fpdf==1.7.2
idna==3.4
img2pdf==0.4.4
iniconfig==1.1.1
install==1.3.5
kiwisolver==1.4.2
lxml==4.9.1
matplotlib==3.5.2
mock==4.0.3
MouseInfo==0.1.3
mplcursors==0.5.3
mss==9.0.1
numpy==1.22.4
opencv-python==4.6.0.66
opencv-python-headless==4.8.1.78
packaging==21.3
pandas==1.4.2
path==16.4.0
path.py==12.5.0
pdfminer.six==20220524
pdfplumber==0.7.5
pefile==2023.2.7
pikepdf==5.3.2
Pillow==10.0.1
pluggy==1.0.0
py==1.11.0
pyautocad==0.2.0
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
PyMsgBox==1.0.9
PyMuPDF==1.23.4
PyMuPDFb==1.23.3
pynput==1.7.6
pypandoc==1.13
pyparsing==3.0.9
PyPDF2==2.11.1
pyperclip==1.8.2
PyRect==0.2.0
PyRTF3==0.47.5
pyscreenshot==3.1
PyScreeze==0.1.29
pytest==7.1.3
pytest-shutil==1.7.0
pyth==0.6.0
python-dateutil==2.8.2
pythonnet==3.0.3
pytweening==1.0.7
pytz==2022.1
pywin32==304
pywin32-ctypes==0.2.2
pywinauto==0.6.8
requests==2.28.1
six==1.16.0
termcolor==1.1.0
tomli==2.0.1
urllib3==1.26.13
Wand==0.6.10
xlwings==0.27.8
